+++ packages/grid/src/features/standard-table/features/checkboxes/UseTableHeadCheckbox.ts
@@ -4,11 +4,12 @@ import {
   useStandardTableConfig,
   useStandardTableState,
 } from "../../hooks/UseStandardTableConfig";
+import { filterItemsOnEnabledCheckboxes } from "../../util/FilterItemsOnEnabledCheckboxes";
 
 export const useTableHeadCheckbox = <TItem>(
   items: Array<TItem> | undefined
 ) => {
-  const { keyResolver } = useStandardTableConfig();
+  const { keyResolver, checkboxDisabledResolver } = useStandardTableConfig();
   const {
     selectedIds: { selectedIds },
   } = useStandardTableState();
@@ -26,18 +27,25 @@ export const useTableHeadCheckbox = <TItem>(
   const onClickCheckbox = useCallback(() => {
     if (items) {
       if (selectionIsEmpty) {
-        dispatch(setSelectedIds(items.map((item) => keyResolver(item))));
+        dispatch(
+          setSelectedIds(
+            items
+              .filter(filterItemsOnEnabledCheckboxes(checkboxDisabledResolver))
+              .map((item) => keyResolver(item))
+          )
+        );
       } else {
         dispatch(clearSelection());
       }
     }
   }, [
-    selectionIsEmpty,
-    clearSelection,
-    dispatch,
     items,
-    keyResolver,
+    selectionIsEmpty,
+    dispatch,
     setSelectedIds,
+    checkboxDisabledResolver,
+    keyResolver,
+    clearSelection,
   ]);
 
   return {